Convert some require()s into redirected stubs
