+++ packages/lesswrong/lib/vulcan-lib/fragments.ts
@@ -1,10 +1,8 @@
 import type { DocumentNode } from 'graphql';
 import gql from 'graphql-tag';
 import * as _ from 'underscore';
-
-// This is safe as it uses `import type`
-// eslint-disable-next-line import/no-restricted-paths
-import type SqlFragment from '@/server/sql/SqlFragment';
+// This has a stub for the client bundle
+import SqlFragment from '@/server/sql/SqlFragment';
 
 interface FragmentDefinition {
   fragmentText: string
@@ -37,7 +35,7 @@ export const registerFragment = (fragmentTextSource: string): void => {
 
   const sqlFragment = bundleIsServer
     // eslint-disable-next-line import/no-restricted-paths, babel/new-cap
-    ? new (require("@/server/sql/SqlFragment").default)(
+    ? new SqlFragment(
       fragmentText,
       (name: FragmentName) => Fragments[name].sqlFragment ?? null,
     )
