+++ packages/lesswrong/server/commandLine.ts
@@ -1,6 +1,14 @@
-import { CommandLineArguments, isAnyTest, isMigrations } from '../lib/executionEnvironment';
+import { isAnyTest, isMigrations } from '../lib/executionEnvironment';
 import process from 'process';
 import fs from 'fs';
+
+export interface CommandLineArguments {
+  postgresUrl: string
+  postgresReadUrl: string
+  settingsFileName: string
+  shellMode: boolean,
+  command?: string,
+}
 
 
 const parseCommandLine = (argv: Array<string>): CommandLineArguments => {
