+++ packages/lesswrong/server/serverMain.ts
@@ -15,7 +15,7 @@ import { addLegacyRssRoutes } from './legacy-redirects/routes';
 import { initReviewWinnerCache } from './resolvers/reviewWinnerResolvers';
 import { startAnalyticsWriter } from './analyticsWriter';
 import { startSyncedCron } from './cronUtil';
-import { isAnyTest, isMigrations, CommandLineArguments } from '@/lib/executionEnvironment';
+import { isAnyTest, isMigrations } from '@/lib/executionEnvironment';
 import { Globals, Vulcan } from '@/lib/vulcan-lib/config';
 import chokidar from 'chokidar';
 import fs from 'fs';
@@ -24,6 +24,7 @@ import { initGatherTownCron } from './gatherTownCron';
 import { registerViewCronJobs } from './postgresView';
 import { addCountOfReferenceCallbacks } from './callbacks/countOfReferenceCallbacks';
 import { registerElasticCallbacks } from './search/elastic/elasticCallbacks';
+import type { CommandLineArguments } from './commandLine';
 
 /**
  * Entry point for the server, assuming it's a webserver (ie not cluster mode,
