+++ packages/lesswrong/server/serverStartup.ts
@@ -4,10 +4,10 @@ import { createSqlConnection } from './sqlConnection';
 import { replaceDbNameInPgConnectionString, setSqlClient } from './sql/sqlClient';
 import PgCollection, { DbTarget } from './sql/PgCollection';
 import { Collections } from '../lib/vulcan-lib/getCollection';
-import { isAnyTest, CommandLineArguments } from '../lib/executionEnvironment';
+import { isAnyTest } from '../lib/executionEnvironment';
 import { PublicInstanceSetting } from "../lib/instanceSettings";
 import { refreshSettingsCaches } from './loadDatabaseSettings';
-import { getCommandLineArguments } from './commandLine';
+import { CommandLineArguments, getCommandLineArguments } from './commandLine';
 import { Globals } from '../lib/vulcan-lib/config';
 import { getBranchDbName } from "./branchDb";
 import { dropAndCreatePg } from './testingSqlClient';
