+++ packages/lesswrong/stubs/server/commandLine.ts
@@ -1,0 +1,6 @@
+
+export type CommandLineArguments = {
+}
+
+export function loadInstanceSettings() {
+}
