+++ packages/lesswrong/stubs/server/sql/SqlFragment.ts
@@ -1,0 +1,1 @@
+export default null;
